Refactor add store the access token in the cache
